Changed a float to an int & adjusted comments

+++ jc_13c_frame.h
@@ -24,8 +24,8 @@ class CAN_frame_13c : public Converted_CAN_Data
 private:
 	static canid_t frame_id;
 
-    float throttle_command;
-	float throttle_plate;
+    int throttle_command;
+	int throttle_plate;
 	float load_command;
 
 	bool clutch_status;
@@ -34,12 +34,10 @@ public:
 	canid_t get_class_ID(void);
     void convert_frame(struct can_frame &);
 
-    float get_throttle_command(void);
-    float get_throttle_plate(void);
+    int get_throttle_command(void);
+    int get_throttle_plate(void);
     float get_load_command(void);
     bool get_clutch_status(void);
-    
-    //void print_test(void);
 };
 
 #endif
